tela de cadastro de cartao e melhorias

+++ app/src/main/java/com/example/picpaytest/data/model/User.kt
@@ -3,7 +3,7 @@ package com.example.picpaytest.data.model
 import java.io.Serializable
 
 data class User(
-    val id: Long,
+    val id: Int,
     val name: String,
     val img: String,
-    val username: String) : Serializable
+    val username: String): Serializable
